glucose-BE/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
requests
flask
django-cors-headers
google-auth
google-auth-oauthlib
ollama
pillow 
torch 
torchvision
pandas
joblib
scikit-learn